Move unit tests to new code location
